+++ src/app/app.ts
@@ -21,33 +21,16 @@ import apiKeys from './service/constants/apiKeys';
 
 initializeSentryIO();
 const client: Client = initializeClient();
-initializeEvents();
+initializeDiscordEvents();
 
 const creator = new SlashCreator({
 	applicationID: process.env.DISCORD_BOT_APPLICATION_ID,
 	publicKey: process.env.DISCORD_BOT_PUBLIC_KEY,
 	token: process.env.DISCORD_BOT_TOKEN,
+	disableTimeouts: true,
 });
 
-creator.on('debug', (message) => Log.debug(`debug: ${ message }`));
-creator.on('warn', (message) => Log.warn(`warn: ${ message }`));
-creator.on('error', (error: Error) => Log.error(`error: ${ error }`));
-creator.on('synced', () => Log.debug('Commands synced!'));
-creator.on('commandRegister', (command: SlashCommand) => Log.debug(`Registered command ${command.commandName}`));
-creator.on('commandError', (command: SlashCommand, error: Error) => Log.error(`Command ${command.commandName}:`, {
-	indexMeta: true,
-	meta: {
-		name: error.name,
-		message: error.message,
-		stack: error.stack,
-		command,
-	},
-}));
-
-// Ran after the command has completed
-creator.on('commandRun', (command:SlashCommand, result: Promise<any>, ctx: CommandContext) => {
-	LogUtils.logCommandEnd(ctx);
-});
+initializeSlashCreateEvents();
 
 // Register command handlers
 creator
@@ -56,9 +39,6 @@ creator
 	)
 	.registerCommandsIn(path.join(__dirname, 'commands'))
 	.syncCommands();
-
-// Log client errors
-client.on('error', Log.error);
 
 client.login(process.env.DISCORD_BOT_TOKEN).catch(Log.error);
 
@@ -72,11 +52,9 @@ function initializeClient(): Client {
 			Intents.FLAGS.GUILD_VOICE_STATES,
 			Intents.FLAGS.GUILD_PRESENCES,
 			Intents.FLAGS.GUILD_MESSAGES,
-			Intents.FLAGS.GUILD_MESSAGE_REACTIONS,
 			Intents.FLAGS.DIRECT_MESSAGES,
-			Intents.FLAGS.DIRECT_MESSAGE_REACTIONS,
 		],
-		partials: ['MESSAGE', 'CHANNEL', 'REACTION', 'USER'],
+		partials: ['MESSAGE', 'CHANNEL', 'USER'],
 	};
 	return new Discord.Client(clientOptions);
 }
@@ -96,18 +74,69 @@ function initializeSentryIO() {
 	});
 }
 
-function initializeEvents(): void {
-	const eventFiles = fs.readdirSync(path.join(__dirname, '/events')).filter(file => file.endsWith('.js'));
+function initializeDiscordEvents(): void {
+	// Log client errors
+	client.on('error', Log.error);
+
+	// register other events
+	const eventFiles = fs.readdirSync(path.join(__dirname, '/events/discordjs')).filter(file => file.endsWith('.js'));
 	eventFiles.forEach(file => {
-		const event = new (require(`./events/${file}`).default)();
+		const event = new (require(`./events/discordjs/${file}`).default)();
 		try {
 			if (event.once) {
 				client.once(event.name, (...args) => event.execute(...args, client));
 			} else {
-				client.on(event.name, (...args) => event.execute(...args, client));
+				client.on(event.name, (...args) => {
+					event.execute(...args, client);
+				});
 			}
+			Log.debug(`registered discordjs event ${event.name}`);
 		} catch (e) {
 			Log.error('Event failed to process', {
+				indexMeta: true,
+				meta: {
+					name: e.name,
+					message: e.message,
+					stack: e.stack,
+					event,
+				},
+			});
+		}
+	});
+}
+
+function initializeSlashCreateEvents(): void {
+	creator.on('debug', (message) => Log.debug(`debug: ${ message }`));
+	creator.on('warn', (message) => Log.warn(`warn: ${ message }`));
+	creator.on('error', (error: Error) => Log.error(`error: ${ error }`));
+	creator.on('synced', () => Log.debug('Commands synced!'));
+	creator.on('commandRegister', (command: SlashCommand) => Log.debug(`Registered command ${command.commandName}`));
+	creator.on('commandError', (command: SlashCommand, error: Error) => Log.error(`Command ${command.commandName}:`, {
+		indexMeta: true,
+		meta: {
+			name: error.name,
+			message: error.message,
+			stack: error.stack,
+			command,
+		},
+	}));
+	// Ran after the command has completed
+	creator.on('commandRun', (command:SlashCommand, result: Promise<any>, ctx: CommandContext) => {
+		LogUtils.logCommandEnd(ctx);
+	});
+
+	const eventFiles = fs.readdirSync(path.join(__dirname, '/events/slash-create')).filter(file => file.endsWith('.js'));
+	eventFiles.forEach(file => {
+		const event = new (require(`./events/slash-create/${file}`).default)();
+		try {
+			if (event.once) {
+				creator.once(event.name, (...args) => event.execute(...args, client));
+			} else {
+				creator.on(event.name, (...args) => event.execute(...args, client));
+			}
+			Log.debug(`Registered slash-create event ${event.name}`);
+		} catch (e) {
+			Log.error('Slash-create event failed to process', {
 				indexMeta: true,
 				meta: {
 					name: e.name,